test(solver): move keyof display-alias test to a dedicated shard
Address review: keep the new formatter test out of the already-oversized
`diagnostics/format/tests.rs` (file-size hard gate). Move
`keyof_display_alias_does_not_repaint_unit_literal_union` into a new
`diagnostics/format/keyof_alias_display_tests.rs` shard and wire it into the
module; `tests.rs` returns to its prior length.
